Consolidate remaining fabric source files, move utils
