+++ src/sources/image.ts
@@ -1,9 +1,8 @@
 import { FabricImage, type StaticCanvas } from 'fabric/node';
 import { blurImage } from './fabric.js';
-import { getZoomParams, getTranslationParams } from './fabricFrameSources.js';
-import { loadImage } from '../util.js';
-import type { ImageLayer } from '../types.js';
+import { getZoomParams, getTranslationParams, loadImage } from '../util.js';
 import { defineFrameSource } from './index.js';
+import type { ImageLayer } from '../types.js';
 
 export default defineFrameSource<ImageLayer>(async ({ verbose, params, width, height }) => {
   const { path, zoomDirection = 'in', zoomAmount = 0.1, resizeMode = 'contain-blur' } = params;
