+++ src/sources/news-title.ts
@@ -1,7 +1,7 @@
 import { Rect, FabricText } from 'fabric/node';
 import { easeOutExpo } from '../transitions.js';
 import type { NewsTitleLayer } from '../types.js';
-import { defaultFontFamily } from './fabricFrameSources.js';
+import { defaultFontFamily } from '../util.js';
 import { defineFrameSource } from './index.js';
 
 export default defineFrameSource<NewsTitleLayer>(async ({ width, height, params }) => {
