+++ src/sources/radial-gradient.ts
@@ -1,7 +1,7 @@
 import * as fabric from 'fabric/node';
 import { getRandomGradient } from '../colors.js';
 import type { RadialGradientLayer } from '../types.js';
-import { getRekt } from './fabricFrameSources.js';
+import { getRekt } from '../util.js';
 import { defineFrameSource } from './index.js';
 
 export default defineFrameSource<RadialGradientLayer>(async ({ width, height, params }) => {
